QP: Refactor statements into packages.
