+++ test/QueryProcessorTests/ParserTests.java
@@ -1,13 +1,15 @@
 package QueryProcessorTests;
 
 import Predicate.*;
-import QueryParser.DDLStatement.*;
-import QueryParser.DMLStatement.*;
-import QueryParser.InternalStatement.*;
-import QueryParser.Parser;
-import QueryParser.PredicateParser.*;
-import QueryParser.SQLStatement;
-import QueryParser.TokenKind;
+import QueryParser.*;
+import QueryParser.DDLStatements.CreateIndexStatement;
+import QueryParser.DDLStatements.CreateTableStatement;
+import QueryParser.DDLStatements.DDLStatement.*;
+import QueryParser.DMLStatements.*;
+import QueryParser.DMLStatements.DMLStatement.*;
+import QueryParser.InternalStatements.CompactStatement;
+import QueryParser.InternalStatements.DescribeStatement;
+import QueryParser.InternalStatements.InternalStatement.*;
 import RheaDB.Attribute;
 import RheaDB.AttributeType;
 import RheaDB.DBError;
@@ -255,7 +257,7 @@ public class ParserTests {
     }
 
     @Test
-    void parseInvalidUpdateWithoutWherePredicates() throws DBError {
+    void parseInvalidUpdateWithoutWherePredicates() {
         String sqlString = "UPDATE FancyTable SET intAttribute = 1, stringAttribute = "
                 + "\"Random String\" WHERE";
 
@@ -287,74 +289,5 @@ public class ParserTests {
 
         var updateStatement = testInitialPartOfUpdate(sqlString);
         Assertions.assertTrue(updateStatement.getWherePredicates().isEmpty());
-    }
-
-    @Test
-    void parseCompoundPredicate() throws Exception {
-        String predicate = "(a > 10 && b <= 12) || a != 2";
-        PredicateAST ast = new PredicateParser(predicate).parse();
-        Assertions.assertTrue(ast.root() instanceof BinaryLogicalExpression);
-
-        var left = ((BinaryLogicalExpression) ast.root()).getLhs();
-        var operator = ((BinaryLogicalExpression) ast.root()).getOperatorToken();
-        var right = ((BinaryLogicalExpression) ast.root()).getRhs();
-
-        Assertions.assertTrue(left instanceof ParenthesizedExpression);
-        Assertions.assertEquals(operator.getKind(), TokenKind.BarBarToken);
-        Assertions.assertTrue(right instanceof BinaryLogicalExpression);
-
-        var enclosedExpression = ((ParenthesizedExpression) left).getExpression();
-        Assertions.assertTrue(enclosedExpression instanceof BinaryLogicalExpression);
-
-        left = ((BinaryLogicalExpression) enclosedExpression).getLhs();
-        operator = ((BinaryLogicalExpression) enclosedExpression).getOperatorToken();
-        right = ((BinaryLogicalExpression) enclosedExpression).getRhs();
-
-        Assertions.assertTrue(left instanceof BinaryLogicalExpression);
-        Assertions.assertEquals(operator.getKind(), TokenKind.AmpersandAmpersandToken);
-        Assertions.assertTrue(right instanceof BinaryLogicalExpression);
-
-        var aGreaterThan10 = (BinaryLogicalExpression) left;
-        Assertions.assertNotNull(aGreaterThan10);
-        var aIdentifier = aGreaterThan10.getLhs();
-        var greaterOperator = aGreaterThan10.getOperatorToken();
-        var tenLiteral = aGreaterThan10.getRhs();
-
-        Assertions.assertTrue(aIdentifier instanceof IdentifierExpression);
-        Assertions.assertEquals(((IdentifierExpression) aIdentifier).getIdentifierToken().getValue(), "a");
-        Assertions.assertEquals(greaterOperator.getKind(), TokenKind.GreaterToken);
-        Assertions.assertTrue(tenLiteral instanceof LiteralExpression);
-        Assertions.assertEquals(((LiteralExpression) tenLiteral).getValueToken().getValue(), 10);
-
-        var bLessEqualsThan12 = (BinaryLogicalExpression) right;
-        Assertions.assertNotNull(bLessEqualsThan12);
-
-        var bIdentifier = bLessEqualsThan12.getLhs();
-        var lessEquals = bLessEqualsThan12.getOperatorToken();
-        var twelveLiteral = bLessEqualsThan12.getRhs();
-
-        Assertions.assertTrue(bIdentifier instanceof IdentifierExpression);
-        Assertions.assertEquals(((IdentifierExpression) bIdentifier).getIdentifierToken().getValue(), "b");
-        Assertions.assertEquals(lessEquals.getKind(), TokenKind.LessEqualsToken);
-        Assertions.assertTrue(twelveLiteral instanceof LiteralExpression);
-        Assertions.assertEquals(((LiteralExpression) twelveLiteral).getValueToken().getValue(), 12);
-    }
-
-    @Test
-    void testIdentifierParsesAsNullPredicate() throws Exception {
-        var result = new PredicateParser("identifierName").parse();
-        Assertions.assertNull(result);
-    }
-
-    @Test
-    void testLiteralsParseAsNullPredicate() throws Exception {
-        var result = new PredicateParser("123").parse();
-        Assertions.assertNull(result);
-
-        result = new PredicateParser("\"HELLO\"").parse();
-        Assertions.assertNull(result);
-
-        result = new PredicateParser("3.1415296").parse();
-        Assertions.assertNull(result);
     }
 }
